Remove phase_c_exports.rs link-bridge; relocate its live exports
phase_c_exports.rs was a transient link bridge from the Rust port that had
accumulated a grab-bag of C-ABI exports plus hundreds of comments tracking
where each symbol had moved. Distribute the still-live exports to their
proper homes and delete the file.

- `Bun__panic`, `JSC__JSValue__parseJSON`, `Bun__LifecycleAgentPreventExit`,
  `Bun__LifecycleAgentStopPreventingExit` move to the binary crate root
  (`src/bun_bin/lib.rs`), next to `main`/`__asan_default_options`. These are
  symbols the final link must provide: `bun_jsc` declares and calls them but
  C++ has no body, and the declaring crate cannot also define the symbol.

- `Bun__VM__scriptExecutionStatus` moves to `src/jsc/VirtualMachine.rs` as an
  `export_name` wrapper over the existing `script_execution_status()` method,
  matching the sibling `Bun__VM__*` exports there. It was a placeholder that
  hardcoded `0` (Running); the method returns the real status, and the C++
  caller already switches on all three values.

- `DNSResolver__getConstructor` needed no relocation. The `#[JsClass]` derive
  emits that extern unless the attribute carries `no_constructor`; DNSResolver
  is `noConstructor: true` in its `.classes.ts` but the flag was missing, so
  the stub existed to satisfy the dangling reference. Add the flag (as
  NodeJSFS already does) and the macro falls back to the trait's default
  `get_constructor` returning `undefined` — no stub, no extern.

- `BunString__toErrorInstance` is deleted outright: nothing references it.

The Rust side already consulted `script_execution_status()` directly
everywhere, so only the C++/JSC export changes, and there is no teardown path
where JS can observe the difference.
